Update vote extensions
